fix: guard MIDTOWN_SESSION_ID injection to stable IDs only, move tests to launch_tests.rs
- Only inject env var for Claude sessions (or resumed sessions) — fresh
  Codex sessions use a provisional UUID that is replaced on init, so
  injecting it would leave a stale $MIDTOWN_SESSION_ID after migration
- Move inject_session_id_env unit tests from inline mod tests block in
  launch.rs to launch_tests.rs per repo convention
